2004-05-18  Pierre-Luc Paour <gallery@paour.com> (1.4-b13)
	* Added support for non-8bit characters in album and image data.
	  If GR is running in a non-international system and the default
	  font used by Java is not able to display the text correctly,
	  the FONT_OVERRIDE_* properties can force GR to use any system
	  font.
	  This only works if the album and image data are encoded by
	  Gallery in HTML entities (&#xxxx;). When adding albums and
	  images, it only works if automatic HTML escaping is on.
	* Added support for running GR in l10n-development mode by setting
	  uiLocaleDev=true in the prefs. This will load the language files
	  from img/ so they're easy to modify.
	* Fixed a problem in the applet where using the applet and then
	  moving to another tab and coming back would result in the list
	  of pictures being "Problem...".
	* Added credits for translators in Locale prefs.
	* Fixed some bugs when no Gallery profile yet exists.

+++ com/gallery/GalleryRemote/util/ImageUtils.java
@@ -370,6 +370,12 @@ public class ImageUtils {
 				toDelete.add(r);
 
 				ImageWriter writer = ImageIO.getImageWriter(reader);
+
+				if (writer == null) {
+					Log.log(Log.LEVEL_ERROR, MODULE, "No writer to write out " + filename + " ImageIO probably doesn't support it. Resize aborted.");
+					return new File(filename);
+				}
+
 				ImageOutputStream ios = null;
 				try {
 					r.delete();
@@ -399,6 +405,7 @@ public class ImageUtils {
 			}
 		} catch (IOException e) {
 			Log.logException(Log.LEVEL_ERROR, MODULE, e);
+			return new File(filename);
 		}
 
 		return r;
